move region_check test from raftstore to tikv_util
