+++ components/tikv_util/src/store/region.rs
@@ -37,3 +37,35 @@ pub fn region_on_same_stores(lhs: &Region, rhs: &Region) -> bool {
             .any(|rp| rp.get_store_id() == lp.get_store_id() && rp.get_role() == lp.get_role())
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    // Tests the util function `check_key_in_region`.
+    #[test]
+    fn test_check_key_in_region() {
+        let test_cases = vec![
+            ("", "", "", true, true, false),
+            ("", "", "6", true, true, false),
+            ("", "3", "6", false, false, false),
+            ("4", "3", "6", true, true, true),
+            ("4", "3", "", true, true, true),
+            ("3", "3", "", true, true, false),
+            ("2", "3", "6", false, false, false),
+            ("", "3", "6", false, false, false),
+            ("", "3", "", false, false, false),
+            ("6", "3", "6", false, true, false),
+        ];
+        for (key, start_key, end_key, is_in_region, inclusive, exclusive) in test_cases {
+            let mut region = Region::default();
+            region.set_start_key(start_key.as_bytes().to_vec());
+            region.set_end_key(end_key.as_bytes().to_vec());
+            let mut result = check_key_in_region(key.as_bytes(), &region);
+            assert_eq!(result, is_in_region);
+            result = check_key_in_region_inclusive(key.as_bytes(), &region);
+            assert_eq!(result, inclusive);
+            result = check_key_in_region_exclusive(key.as_bytes(), &region);
+            assert_eq!(result, exclusive);
+        }
+    }
+}
